Renaming library from 'git' to 'scm' to avoid naming conflicts.
