+++ kernel/scm/git/__init__.py
