+++ kernel/scm/git/apis/utils.py
