+++ tests/kernel/intr/scm/git/apis/cli/test_gitpython.py
@@ -16,7 +16,7 @@
 from tempfile import TemporaryDirectory
 from unittest import TestCase
 
-from kernel.git.apis.cli.gitpython import GitPython
+from kernel.scm.git.apis.cli.gitpython import GitPython
 from kernel.tools.fs import Dir
 from kernel.tools.urls import URL
