+++ tests/kernel/intr/scm/git/apis/test_utils.py
@@ -15,7 +15,7 @@
 """
 from unittest import TestCase
 
-from kernel.git.apis.utils import get_repository_fullname
+from kernel.scm.git.apis.utils import get_repository_fullname
 from kernel.tools.urls import URL
