+++ tripleo/insights/git.py
@@ -20,13 +20,13 @@ from typing import Iterable, Optional, Sequence
 from cached_property import cached_property
 from overrides import overrides
 
-from kernel.git.apis.cli import Git, GitError
-from kernel.git.apis.cli import Repository as GitRepo
-from kernel.git.apis.cli.gitpython import GitPython
-from kernel.git.apis.github import GitHub, GitHubError
-from kernel.git.apis.github import Repository as GitHubRepo
-from kernel.git.apis.github.pygithub import PyGitHub
-from kernel.git.apis.utils import get_repository_fullname
+from kernel.scm.git.apis.cli import Git, GitError
+from kernel.scm.git.apis.cli import Repository as GitRepo
+from kernel.scm.git.apis.cli.gitpython import GitPython
+from kernel.scm.git.apis.github import GitHub, GitHubError
+from kernel.scm.git.apis.github import Repository as GitHubRepo
+from kernel.scm.git.apis.github.pygithub import PyGitHub
+from kernel.scm.git.apis.utils import get_repository_fullname
 from kernel.tools.fs import Dir
 from kernel.tools.paths import resolve_home
 from kernel.tools.rng import get_new_uuid
